requirements.txt
asgiref==3.4.1
attrs==21.2.0
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.8
django-environ==0.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.0
Faker==9.5.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==2.3.0
pyparsing==3.0.0
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
